Renamed data package where we store model classes
