+++ data/src/main/java/com/example/data/source/LocationDataSource.kt
@@ -1,6 +1,6 @@
 package com.example.data.source
 
-import com.example.data.entities.Location
+import com.example.data.model.Location
 import kotlinx.coroutines.flow.Flow
 
 interface LocationDataSource {
